requirements.txt
streamlit>=1.28.0
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-community>=0.0.20
faiss-cpu>=1.8.0
python-dotenv>=1.0.0
plotly>=5.15.0
streamlit-lottie>=0.0.5
streamlit-option-menu>=0.3.6
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0
textblob>=0.17.0
tiktoken>=0.5.0
pydantic>=2.0.0